test: add jest/supertest infrastructure and backend specs
- restructure server bootstrap to export the Express app and share config with tests
- configure Jest (ESM-friendly) with a dedicated .env.test loader and health check integration test
- mock Supabase auth to cover /api/auth/login success and failure scenarios
- document npm test usage and required test environment
