+++ backend/.env.test
@@ -1,0 +1,8 @@
+SUPABASE_URL=http://localhost:54321/test-project
+SUPABASE_KEY=test-anon-key
+SUPABASE_SERVICE_ROLE_KEY=test-service-role-key
+OPENAI_API_KEY=dummy-test-key
+ALLOWED_ORIGINS=http://localhost:5173
+APP_NAME=Kadi (test)
+APP_URL=http://localhost:5173
+EMAIL_REDIRECT_URL=http://localhost:5173/login?confirmed=1
