+++ backend/jest.config.js
@@ -1,0 +1,7 @@
+export default {
+  testEnvironment: 'node',
+  transform: {},
+  moduleFileExtensions: ['js', 'mjs'],
+  roots: ['<rootDir>/tests'],
+  setupFiles: ['<rootDir>/tests/setupEnv.js']
+}
