+++ backend/package.json
@@ -4,7 +4,8 @@
   "type": "module",
   "scripts": {
     "dev": "nodemon server.js",
-    "start": "node server.js"
+    "start": "node server.js",
+    "test": "cross-env NODE_ENV=test NODE_OPTIONS=--experimental-vm-modules jest --runInBand"
   },
   "dependencies": {
     "@supabase/supabase-js": "^2.45.2",
@@ -18,6 +19,9 @@
     "pdfkit": "^0.15.0"
   },
   "devDependencies": {
-    "nodemon": "^3.1.7"
+    "cross-env": "^7.0.3",
+    "jest": "^29.7.0",
+    "nodemon": "^3.1.7",
+    "supertest": "^6.3.4"
   }
 }
