+++ backend/tests/setupEnv.js
@@ -1,0 +1,3 @@
+import dotenv from 'dotenv'
+
+dotenv.config({ path: '.env.test', override: true })
